Chris_solution/DIC_Ex3_Valentin/Our results/requirements (1).txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
fastapi
tflite
tensorflow
tflite-runtime
tensorflow_hub
